3.4 fixed, 3.8 started

+++ chapter3/task4/main.cpp
@@ -2,9 +2,9 @@
 
 #include <iostream>
 int main() {
-  long a;
-  long b;
-  long c;
+  float a;
+  float b;
+  float c;
   char ch;
   char next = 'y';
   while (next == 'y'){
